% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMBH.R
\name{fitMBH}
\alias{fitMBH}
\title{fitMBH}
\usage{
fitMBH(x, vars = c("V1", "V2", "V3"), groups = "Group", nc = 3,
  ni = 1e+05, nb = 20000, nt = 20)
}
\arguments{
\item{x}{Data to fit hypervolume to}

\item{vars}{Names of variables in x to use for hypervolume construction}

\item{groups}{Name of the grouping variable in x. Use NULL if no groups present or to  ignore grouping structure and fit an empirical hypervolume}

\item{nc}{Number of MCMC chains}

\item{ni}{Number of MCMC iterations (default 100000)}

\item{nb}{Length of burnin}

\item{nt}{Thinning paramter}
}
\value{
means - Estimated means of each variable

covariance - Estimated covariance structure

volume - Estimated hypervolume size

group_means - Estimated group means

group_variances - Estimated between-group variances for each variable

samples - The output from the jags.samples function
}
\description{
This function fits modelled based or empirical hypervolumes to multivariate data
}
\details{
To use the coda package for mcmc diagnostics, you first need to convert the samples to mcmc.list format. This can be completed with the as.mcmc.list function from the mcmcr package. To inspect the mcmc chains for the estimated covariance matrix use plot(as.mcmc.list(m3$samples$tau))
}
