% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataMBH.R
\name{simulate_dataMBH}
\alias{simulate_dataMBH}
\title{simulate_dataMBH}
\usage{
simulate_dataMBH(nobs = 10, ndims = 3, ngroups = 4, sdgrp = 2,
  sdobs = NULL, variances = "fixed", vardiff = 0.5, means = rep(0,
  ndims), returntruecov = FALSE)
}
\arguments{
\item{nobs}{Number of observations to simulate per group}

\item{ndims}{Number of variables to simulate}

\item{ngroups}{Number of groups to simulate}

\item{sdgrp}{Standard deviation of group means}

\item{sdobs}{Standard deviation of observations within groups. Either a single value or vector with length equal to ndims. If NULL (default) a vector of length equal to ndims is generated from a Uniform distribution with bounds 1 and 10. Note that if sdobs = NULL the vector generated is on the variance scale, if sdobs is provided then it should be a standard deviation}

\item{variances}{Either "fixed" to constrain within-group variances to be the same (the default), or "variable" so within-group variances can vary}

\item{vardiff}{Standard deviation of within-group variance differences}

\item{means}{Means of each variable to simulate}

\item{returntruecov}{Logical. Return true covariance matrix as well as data? Defaults to FALSE}
}
\value{
data - simulated data

(optional) truecov - covariance matrix used to simulate data
}
\description{
This simulates multivariate data to fit model based hypervolumes
}
