% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclMBH.R
\name{inclMBH}
\alias{inclMBH}
\title{inclMBH}
\usage{
inclMBH(hv, newdat)
}
\arguments{
\item{hv}{Fitted MBH model}

\item{newdat}{New data to test probability of inclusion in the hypervolume}
}
\value{
Returns newdat with an additional column corresponding to probability of inclusion in the hypervolume
}
\description{
This calculates the probability of inclusion of new data into a calculated hypervolume
}
